改表ID字段为Int

+++ src/main/kotlin/com/linweiyuan/mhp/entity/User.kt
@@ -9,7 +9,7 @@ data class User(
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         @JsonIgnore
-        val id: Byte,
+        val id: Int,
         val username: String,
         val password: String,
         @JsonIgnore
